requirements.txt
# Dependencias del proyecto Wellness
fastapi
uvicorn[standard]
python-jose[cryptography]
bcrypt
scikit-learn
pandas
numpy
matplotlib